Update problem S3_elem_3_008540 for localization and layout adjustments
- Adjusted diagram label positions and text formatting in Korean, Ukrainian, and Chinese versions of the problem.
- Updated semantic roles for various elements to improve accessibility and clarity.
- Modified SVG representations to reflect the new positions and styles.
- Ensured consistency in font usage across different language versions.
- Corrected paths in test files to point to the appropriate storage locations for problem assets.

+++ src/modu_math/renderer/compiler.py
@@ -357,7 +357,7 @@ def _compile_slots(
             tx = float(content["x"]) if isinstance(content.get("x"), int | float) else x
             ty = float(content["y"]) if isinstance(content.get("y"), int | float) else y
             font_size = (
-                int(content["font_size"])
+                float(content["font_size"])
                 if isinstance(content.get("font_size"), int | float)
                 else _DEFAULT_FONT_SIZE
             )
